[feature][language]: factorization of controllers
