+++ config/services.yaml
@@ -37,6 +37,8 @@ services:
             tags: ['app.event.interface']
         App\Interfaces\Event\ViewEventInterface:
             tags: ['app.view.event.interface']
+        App\Interfaces\Factory\EntityFactoryInterface:
+            tags: ['app.entity.factory.interface']
 
     App\Service\EventService:
         # inject all services tagged with app.event.interface as first argument
@@ -47,3 +49,9 @@ services:
     App\Service\ViewService:
         arguments:
             - !tagged_iterator app.view.event.interface
+
+    App\Service\EntityService:
+        arguments:
+            - !tagged_iterator app.entity.factory.interface
+            - '@doctrine.orm.default_entity_manager'
+            - '@App\Service\UuidService'
