+++ src/Interfaces/Factory/EntityFactoryInterface.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace App\Interfaces\Factory;
+
+interface EntityFactoryInterface
+{
+    /**
+     * @return string
+     */
+    public function getRelatedEntity(): string;
+}
